user id and task status

+++ app/Http/Controllers/TaskController.php
@@ -11,35 +11,64 @@ class TaskController extends Controller
 
     public function index()
     {
-        $response = Task::where('user_id', auth()->id())->latest()->get();
+        $user = auth()->user();
+        $response = Task::where('user_id', $user->id)->latest()->get();
         return response()->json($response, 200);
     }
 
     public function store(TaskRequest $request)
     {
+        $user = auth()->user();
         $data = $request->validated();
-        $data['user_id'] = auth()->id();
+        $data['user_id'] = $user->id;
         $response = Task::create($data);
         return response()->json($response, 201);
     }
 
     public function show(Task $task)
     {
-        $response = $task;
-        return response()->json($response, 200);
+        $user = auth()->user();
+        if ($user->id == $task->user_id) {
+            $response = $task;
+            return response()->json($response, 200);
+        }else {
+            abort(403);
+        }
     }
 
     public function update(TaskRequest $request, Task $task)
     {
-        $data = $request->validated();
-        $task->update($data);
-        $response = $task;
-        return response()->json($response, 200);
+        $user = auth()->user();
+        if ($user->id == $task->user_id) {
+            $data = $request->validated();
+            $task->update($data);
+            $response = $task;
+            return response()->json($response, 200);
+        }else {
+            abort(403);
+        }
+    }
+
+    public function changeStatus(Task $task, Request $request)
+    {
+        $user = auth()->user();
+        if ($user->id == $task->user_id) {
+            $task->done = $request->done;
+            $rask->save();
+            return response(null, 200);
+        }else {
+            abort(403);
+        }
     }
 
     public function destroy(Task $task)
     {
-        $task->delete();
-        return response(null, 204);
+        $user = auth()->user();
+        if ($user->id == $task->user_id) {
+            $task->delete();
+            return response(null, 204);
+        }else {
+            abort(403);
+        }
     }
 }
